旧グラフ→新グラフ、への移行に伴う諸々の変更
・GraphView.swiftを削除
・AnotherGraphView.swift -> GraphView.swiftへ改名
・上記の変更に伴い、AnalyticsVCを修正
・GraphViewに使用されていたExtensionを削除
